requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
mss>=9.0.1
Pillow>=10.1.0
pynput>=1.7.6
python-multipart>=0.0.6
pyautogui>=0.9.54
keyboard>=0.13.5
pywin32>=306; sys_platform == "win32"

